+++ tests/unit/ValidationForm.vue
@@ -5,6 +5,7 @@
     :resolver="resolver"
     @submit="onSubmit"
     @dirty="$emit('dirty', $event)"
+    @change="$emit('change', $event)"
   >
     <template
       #default="{
@@ -15,7 +16,8 @@
         dirty: formDirty,
         pristine: formPristine,
         invalid: formInvalid,
-        errors: formErrors
+        errors: formErrors,
+        setError
       }"
     >
       <form @submit.prevent="handleSubmit">
@@ -136,10 +138,14 @@
         </ValidationFieldArray>
 
         <ValidationErrors name="common">
-          <template #default="{ errors }">
-            <BaseErrors :errors="errors" />
+          <template #default="{ errors, submitted }">
+            <BaseErrors :errors="errors" :submitted="submitted" />
           </template>
         </ValidationErrors>
+
+        <button id="setError" type="button" @click="setError('common', { message: 'test' })">
+          Set error
+        </button>
 
         <FormInfo
           :values="values"